requirements.txt
sqlalchemy>=1.4
streamlit
streamlit-carousel
plotly
huggingface_hub
PyPDF2
python-docx
